Bug#30735124 INNODB REPLICASET: ADMINAPI TAKES WRONG MEMBER AS HAVING LATEST VIEW/PRIMARY
acquire_primary() was incorrectly reconnecting to an invalidated member
if it happened to be the last member in the replicaset.

This was because the query to get the list of replicaset members was not
considering that view_id for invalidated member rows will be NULL (which
is converted to 0). This caused the check for newest view_id to fail if
a failover was recorded in the metadata.

Change-Id: I8ef5199ccef841171187ed62125c590f1821bcba