Skip tests which are using the Shell share folder
Tests which are writing to the Shell share folder may cause failures in
other tests if test suite is executed in parallel mode. These tests are
now disabled in parallel runs, and will be executed only in sequential
ones.

In one of the tests, files were using wrong extension, this has been
fixed as well.

Change-Id: I21f50a6fa42a6cef38939a1286f234e92e5b02fa