BUG#30645665 : RS.ADD_INSTANCE(CLONE): CRASH CHECKING STATUS
Fixed uncommon segfault in clone monitoring:

The clone plugin ensures the performance_schema.clone_status table is
cleared out when the clone process starts. However, in some rare and
very specific scenarios a race condition can happen and clone is
considered to be running before actually clearing out the table.

The Shell code to handle the clone monitoring was not considering such
scenario resulting in a segfault.

This patch fixes the issue by:

- Ensuring the clone status is obtained considering the timestamp on
which the clone started. Values are obtained only for clone processes
that started after the timestamp set.

- Ensuring the null pointer causing the segfault is never null no matter
what.

This patch also fixes an infinite loop in the clone monitoring phase
that happens very rarely when clone is extremelly fast. The root cause
of the problem was a wrong resolution used for the timestamp obtained
before clone starts to be used when obtaining the clone_status from P_S.

The resolution of performance_schema.clone_status.begin_time is 3,
regardless of the platform, and the shell code was calculating a
timestamp to compare with using a resolution of 6: NOW(6). This has been
changed to use: NOW(3).

Change-Id: I73aa8da8ff6b18558f392b309a1aa42c7eed6d1d