BUG#30739252 : race condition in forcequorum
The <cluster>.forceQuorumUsingPartitionOf() was not stopping Group
Replication on reachable members that are not part of the visible membership
for the target instance, which could lead to an undefined behavior if any of
those members are auto-rejoining.

This patch fixes this issue by explicitly stopping Group Replication on
reachable members that are not included in the new force quorum membership.

Test added for regression.

Change-Id: I1f673677678f917af8962e0ef7399837bdf00c7d