Bug#30866632,Bug#30657911 Clone restart timeout fixes
Bug#30866632 TIMEOUT FOR SERVER RESTART AFTER CLONE TOO SHORT
- Added a new shell.options["dba.restartWaitTimeout"] with default of 60 to
control how many seconds to wait for the server to come back after clone
restarts it.

Bug#30657911 <REPLICASET>.ADDINSTANCE() USING CLONE SEGFAULTS WHEN RESTART IS NOT AVAILABLE
- Fixed crash in replicaset addInstance() when restart times out after a
clone operation. The addInstance() will now abort and revert changes on
timeout, because the add operation needs to connect to the target
instance to finish its job. The user will have to restart the server and
call addInstance() again to retry. The retry can then use incremental
recovery, which wouldn't trigger the clone and the restart.

Also:
- Reduced restart timeout for slow test
- Merged several tests to avoid re-creating sandboxes
- Fixed a bug that caused a restart monitoring timeout to be multiplied

Change-Id: I54a39fd7ff87109045f0a9fa38aa7e22ba237b5c