Improve upgrade_sandbox, start_sandbox and wait_sandbox_dead from testutils.
This patch disables the mysqlx.so plugin load on testutil.upgradeSandbox.

It also removes the retries on testutil.startSandbox which caused issues on the
cluster_offline_upgrade test on siv14 (because of the upgrade the server was
taking too long to start and it retried launching it again).

Furthermore it introduces a timeout option on testutils.startSandbox which can
be used to control the amount of time in seconds to wait for the sandbox so
start listening for connections.

Finally, it improves the wait_sandbox_method by waiting on doublewrite files
locks to be released.

Change-Id: Id7409202c01c957943d02b84756346ef505730ac