Test Suite Improvements
- Splitting cluster_addinstance.js test
- Improving runtime for cluster options tests
- Adding support for multiline wildcards in multiline validation
- Printing the server error log on sandbox start failures

This patch enables defining chunks of output that should be ignored when
validating multiline output through the use of the ${*} pattern.

Multiline validation ensures the exact presense of the expected lines, in
cases where some output is non deterministic, it should be possible to
ignore such output but still validate the remaining of the expected text.

Using:

Sample multiline output
${*}
End of multiline output

Will cause the output validation to succeed as long as the first and last
lines are found on the actual output, no matter what's in the middle.

Change-Id: Ie351ff590d646f688dd02d2ec79f745b3740d919