Enable wait flag for stop_sandbox call inside testutils.restart_sandbox
With this patch the testutils.restart_sandbox will wait for the server to
properly shutdown and release file locks by enabling the wait flag on the
internal call to stop_sandbox.

Furthermore, this patch also improves the wait_sandbox_dead method by
waiting on `ibtmp1` file locks to be released.

Change-Id: Ic99bfe1a8c28261e5b27c9703b1df6aef8131e90