More reliable server shutdown check in tests
Changed testutil.stopSandbox() to wait for the servers pid file to be
deleted as an indication that the shutdown is done. Previous check based
on TCP port onlywas not reliable and was letting the test continue while
InnoDB was still shutting down, which led to random test failures from
multiple mysqld instances running on the same datadir.

Also removed code to delete pid file after server shutdown in
mysqlprovision, which was unnecessary and prevented proper detection
of the server shutdown.

Change-Id: I9c24f090443abea2c93f65de2b59ea4c613e5112