BUG#30912984 CAN'T CONNECT WITH MYSQL SHELL AFTER CHANGING STORED PASSWORD
When connecting to a server using classic session, if password was
stored in a credential store but later changed, connection failed
without prompting the user for a password.

This bug has been fixed, invalid password is now correctly removed from
the password store and password prompt is presented to the user.

Change-Id: I38c609f6e9cca0b0387f54ae71a85eef27c7cde4