BUG#30825330 CORE DUMP IF AN SQL STATEMENT IS CALLED AFTER A STORED PROCEDURE
When using an X session, if an SQL statement was called after a stored
procedure call whose result was not consumed, an error was generated
and Shell could crash when closing.

Internally, before a new statement is executed, Shell drained data from
the current result set, however it did not check if there are any
additional result sets. Since a stored procedure call always returns
multiple result sets, this caused the data from the second and any
additional result set to be left behind and resulted in error when any
other statement was executed.

This issue is now fixed, the data from the current result set is cached
and available to the user, while any remaining result sets are
discarded.

Change-Id: I541009ef495c46dd191e7ac5f6cf21d65e2d17be