BUG#30728744 MISLEADING ERROR MESSAGE WHEN DIFFERENT CREDENTIALS ARE USED IN CLUSTER ADMIN OP
InnoDB Cluster requires that the cluster administrative user name and
password must be the same on all instances that belong to a cluster.

If the user chooses to use the 'root' account, then it must ensure that all
cluster members have a 'root' account with the same password. If the user
chooses to create a clusterAdmin account using dba.configureInstance(), then
it must ensure that uses the same password in every instance.  However, a
user may not be aware of this and might have set up the accounts using
different credentials.

We test connection to primary (seed) instance using provided credentials and
print error to the user if credentials are invalid.

Change-Id: I50c0310d6c6be0e65e17f307eda8c9f7e69b3a63