BUG#30645697: RECOVERYMETHOD:CLONE WHEN ALL MEMBERS ARE IPV6 WILL ALLOW ERRANT GTIDS
With this patch, the cluster behavior is now similar to ReplicaSet behavior regarding
the handling of IPv6 instances and the clone recovery_method.
If all of the cluster online instances are using IPv6 addresses and the user tries to
use the clone recovery_method, and error is thrown.

Tests were added to verify the new behavior.

Change-Id: I85ade5b7268c43a10bb961c20163b94a24351449