BUG#30749352, BUG#30896573, BUG#30869854, BUG#30863587
This patch fixes several issues:

BUG#30749352: CREATEREPLICASET MENTION GROUP REPLICATION IN ERROR MESSAGE WHEN
USING 127.0.1.1

This issue is fixed by making the existing error message more generic with the
removal of the Group Replication mention. Existing Cluster tests were updated
and new ones were added for replicaSet.

BUG#30896573: OFFLINE_MODE NOT LISTED IN HELPER FOR THE SUPPORTED VALUES OF
EXITSTATEACTION

This patch adds the OFFLINE_MODE option to the helper of the exitStateAction. No
validation is done on the value of OFFLINE_MODE as requested on the bug report
since it was decided that validation of the values would be done in the GR
plugin. Existing help tests were updated.

BUG#30869854: OPTION 'INTERACTIVE' MISSING FROM CLUSTER.ADDINSTANCE HELP

This patch adds the interactive option to the online helper of
cluster.addInstance. It also normalizes all appearances of the interactive
option in all the commands for the cluster and replicaSet objects so that the
help text is the same. Existing help tests were updated.

BUG#30863587: DBA.DELETESANDBOXINSTANCE DOESN'T RETURN ERROR MESSAGE

After this patch, the dba.deleteSandboxInstance command throws a
runtimeError if we try to delete a sandbox that doesn't exist.
However, the testutils.destroySandbox method keeps the same behavior.
A new test was added to exercise the new behavior.

Change-Id: I73f2f528e2041630d338f46e8f647585cc881e34