BUG#30768504: REBOOTCLUSTER PARTIAL FAILURE IF MYSQL 5.7 IF GR PLUGIN UNINSTALLED
The Group Replication sysvars don't exist if the plugin is not loaded, so even
if they're persisted to my.cnf, they won't be loaded unless the GR plugin is
also loaded when the server starts (the server will throw warnings saying that
he doesn't know the GR variables and ignore them). If the GR plugin is installed
afterwards, my.cnf won't be reloaded, so all sysvars will have default values.
With MySQL 8.0 this was not a problem because SET PERSIST is used, it only
affected 5.7.

This patch fixes the way group_replication variables are persisted to
configuration files. Now, group_replication variables are persisted with the
loose_ prefix on the configureLocalInstance command. As a result, once the plugin
is installed they will get the persisted values instead of the default values.

A new test, similar to what was suggested on the bug report, was added to verify
that rebootCluster works as expected and that group_replication variables are
unchanged.

Change-Id: Ic199678053f2e29745ec7f2f2f4241dce581d2d4