!fix: refactor interrupts into scoped idea
This will change behavior of Interrupts and also Scoped_storage.
With this commit the Scoped_storage is thread_local means that Logger,
Interrupts, IConsole, Shell_options and so on need to be always recreated in each thread,
because current_*() may throw in thread without it.
This will also change the way interrups are handled on Windows.
There's extra thread that will take care of calling interrupt().

Change-Id: Ibb18821f587c161b6c1ce77863f79132345dcec6