BUG#31187130 CANCELLING A DUMP WITH A LOW VALUE SET TO MAXRATE OPTION HANGS THE SHELL
When maxRate option was set to a very low value, cancelling the dump
operation using CTRL+C didn't work.

The SIGINT signal is delivered to only one thread in a process, other
threads are not notified and i.e. sleeping threads are not interrupted
automatically. This patch fixes this problem by introducing
implementation of sleep function based on a conditional variable, which
works for multiple threads and cancels waiting in case of CTRL+C.

Change-Id: I31b90b268be30a1835f409c6479be7d21800f342