BUG#31213967 DUMP TO OCI FAIL IF TABLE NAME REQUIRES TO BE ENCODED
The Bucket class uses the Uri_encoder to encode path segments on Object
Storage requests. The default behavior for the Uri_encoder is to honor
percent encoded data, this is, do not re-encode it as it is already valid
for RFC3986.

The dumper does HEX encoding on file names, which causes the Uri_encoder
to simply skip such sequences. The side effect of this is that if the dumper
is writing a file that is meant to be named like:

SAMPLE%20FILE.txt

In OCI such file is created as:

SAMPLE FILE.txt

This caused the dumper to fail when attempting to rename the object because
the rename operation receives the file names in the request body instead of
the path, so the file really never existed as originally intended.

This patch fixes this by adding a boolean flag on the Uri_encoder::encode_path
function to allow deciding if the existing percent encoded data should be skipped
or not (default behavior is yes). The Bucket class will NOT skip already encoded
data so the file names are created exactly as requested by the caller.

Change-Id: I438607ee2798669ced6efee4b577fe80d808c4fd