BUG#30884590: ADDING AN INSTANCE WITH COMPATIBLE GTID SET SHOULDN'T PROMPT FOR CLONE
Adding an instance with a compatible GTID set to a cluster/replicaset on
which provisioning is required should not require any interaction with the
user. Currently, in such scenario, when clone is supported Shell prompts
the user if wants to use clone or abort the operation.

Considering that in such scenario clone is the only possible way to
provision and add the instance to the cluster/replicaset, Shell can assume
clone is safe to be used and proceed with it automatically.

NOTE: instances with empty GTID-set are not considered to have a
compatible GTID-set. Such scenarios are considered to be unsafe/unknown
thus Shell needs to prompt the user on what action is to be taken.

This patch changes that behavior, ensuring no prompt is presented when
clone is the only way to provision and instance and determined to be
safely usable.

Added new regression tests and updated existing tests.

Change-Id: I6474abc78f31e23e20644300a9cfd78b17695b2a