BUG#30896344: CHECK FOR READ-ONLY COMPATIBILITY IN ADDINSTANCE USES WRONG MINIMUM VERSION
The check for instance read-only compatibility, added as part of WL#13084:
AdminAPI: Handling of cross-version policies, is using a wrong MySQL version
as the base version. The cross-version policies were added in Group
Replication in version 8.0.17 and the check is considering that those were
added in 8.0.16.

The outcome is a misleading Warning message indicating that the added instance
is read-only compatible with the cluster, when that's not true (only for
instances 8.0.16).

This patch fixes that by ensuring that the check to verify if an
instance is read-compatible or not with a cluster is only performed if
the target instance has a version >= 8.0.17.

Added regression unit-test

Change-Id: I242027329b67a7b1f998588448023c7bf7bd7297