BUG#29330769: UPDATETOPOLOGYMODE SHOULD DEFAULT TO TRUE
The 'updateTopologyMode' option of <Cluster>.rescan() is unnecessary.

The metadata holds the value configured for the cluster's topology-mode:
single primary-mode or multi-primary mode. However, that information is
just a snapshot since the configuration itself is owned by Group Replication.
Shell's usage of it is simply for validations and informative purposes.

This patch deprecates the option 'updateTopologyMode' and changes the
behavior of <Cluster>.rescan() to always update the topology mode in the
Metadata when a change is detected, informing the user.

Added regression tests and updated existing ones.

Change-Id: If218f8f146dd00892fcbcee9046d9c89b3f52a85