BUG#31018091: CLUSTERADMIN WITH NETMASK IN HOST IS NOT RECOGNIZED
This patch improves the check done on the account that is used to create
the cluster. From now on, accounts that have a netmask are treated
as accounts with wildcards, i.e., we assume that the user knows what
he is doing and skip further checks to verify if the account accepts
remote connections from all instances.

New tests were added to verify that the old behavior has been corrected.

Change-Id: I2320c192151ff0f29e24aaf770e69aae3c6b5f32