Bug#31455419,Bug#31531704,Bug#31632606,Bug#31632554 AdminAPI createCluster, remove, rejoin fixes
Bug #31455419 CREATECLUSTER, ADDINSTANCE LEAVES BAD STAT BEHIND ON ERROR

When createCluster() fails, its metadata record is left behind and GR is
also started. The cluster is left in an inconsistent state that can't be
recovered from when calling createCluster() again, because it's not
possible to cleanly tell if the metadata record belongs to a failed
attempt or an actual cluster.

Metadata changes during createCluster() are now enclosed in a
transaction, so that the both the cluster and instance records are only
committed if the operation succeeds. GR is also stopped if the create
fails.

A repl called __mysql_innodb_cluster_creating_cluster__ is created at the
beginning of createCluster() and deleted when it's successfully done.
This allows interrupted operations to be detected and resumed on retry.

Similarly, addInstance() was changed to ensure that its metadata record
is only inserted when everything else is done. On retry, it will skip
any steps that would cause conflicts because GR is already running.

This introduces a behaviour change, wherein adding an instance that's
part of the group but not in the metadata will just add it to the
metadata, instead of aborting and requesting the user to execute
cluster.rescan()

Bug #31531704 REMOVING UNREACHABLE MEMBER DOESN'T UPDATE GROUP_REPLICATION_GROUP_SEEDS
When an unreachable member is force-removed, the group_seeds syvar was
not updated because the group status was queried from the primary, which
is now gone. The cluster was left in a bad state, because if any of the
instances were restarted, they would not be able to auto-rejoin, because
group_seeds contains an invalid address, which causes GR to abort trying
to join the group.

We now always call the post-removal routine that updates other instances,
even if the removed instance is not reachable. We query the GR endpoint
from the metadata instead of querying the instance.

Bug #31632606 REMOVEINSTANCE() WITH FORCE DOESN'T WORK IF ADDRESS IS NO LONGER RESOLVABLE

removeInstance() fails if the removed instance is not only unreachable
but also unresolvable, which can be the case when the removed instance
is running in a container that takes down its own DNS record when removed
(like in Kubernetes).

Instances couldn't be removed even if force is enabled, because of a
validation that ensured that the given instance is a valid address, by
checking if it's resolveable. If that validation failed, nothing else is
attempted.

The solution was to completely remove address resolution from address
validation not only from removeInstance(), but from the whole AdminAPI.
That check is redundant, because invalid addresses would eventually lead
to an error anyway when a connection is opened to it. Furthermore, the
result of that check could potentially be different from the result of
the actual connection code. IPv6 address syntax validations was left
to remind users that ::1 should be specified as [::1]

Bug #31632554 CLUSTER.REJOININSTANCE() USES WRONG CREDENTIALS

If credentials are not given in the rejoin command, we take them from
the target server's connection options.

Change-Id: Iad686171c21c56123c95d82aa313ef24a3d86560