Bug#30884174,Bug#31695487 and AdminAPI scriptability improvements
Bug #30884174	CLUSTER.STATUS() HANGS FOR TOO LONG WHEN CONNECTION/READ TIMEOUTS HAPPEN

Reduced default connect timeout in AdminAPI from 10s to 2s, so that
operations like cluster.status() don't freeze for a long time
when there are unreachable members.

Added a session member to dba, to return the session used by it.

Bug #31695487	ALLOW DISABLING GROUP_REPLICATION_START_ON_BOOT

In environments where there's an external entity managing the
cluster lifecycle (e.g. Kubernetes), enabling the
group_replication_start_on_boot can conflict with that entity.
To avoid situations where the cluster topology would change
without direct involvement of external orchestrators,
a new startOnBoot option was added to createCluster(), which
will disable that option in the whole cluster.

= Other changes

== Scripting

Added a version property to the shell object.

Added an ErrorCode property to the mysql object, containing
constants with MySQL error codes.

Removed help topic for "plugins" to avoid collision with a
plugin called "plugins".

== AdminAPI

Several cluster exceptions were augmented with error codes
to allow programs that use AdminAPI to perform error handling
specific to each error condition.

Allow dba.checkInstanceConfiguration() function to be executed
even if the instance is already in an InnoDB cluster. This would
allow instance configuration to be verified before or after being
added to the cluster, which can be important in cases like a new
instance restored from backup but with outdated configurations.

Throw an exception in configureInstance() if the restart option is
enabled, but restart failed.

Added groupViewId to the output of cluster.status({extended:1})

=== Added a new mysqlsh Python module

Added a connect_dba() function to the mysqlsh module, to allow
new dba object instances to be created. These dba instances are
independent of the shell global session and can be used from
threads.

== Python

Translate shell exceptions to a new mysqlsh.Error object, which
can be used to query the error code from shcore::Exception objects.

Ensure that the standard __file__ and __cached__ globals are set when
running Python scripts, which are expected by some internal Python
code for things like formatting exceptions.

Added __name__ and __qualname__ attributes to Python methods, which
are standard in Python objects to allow introspection.

Added a isatty() method to stdout/stderr wrappers for Python.

Change-Id: I1acedc3093a120146c0caa1e99028d4563233d72