Bug #31304711	SHELL SHOULD NOT BUFFER RESULTSETS WHEN PRINTING RESULTS
Don't buffer results when printing SQL mode results as table, as this
will cause an OOM or crash when results are too big.

Instead of buffering the whole resultset, only the 1st 1000 rows are
buffered now and used to find column widths. This causes a minor
regression when column values are larger than the 1st 1000 values,
which causes table formatting to misalign for these rows.

Buffering for DevAPI result objects was not changed.

Fixed formatting of bit columns, column width was always the same and
too wide.

Fixed loss of precision bug in ClassicSession.runSql().
placeholder values of double type was being truncated because
std::to_string() was used instead of my_gcvt.

Wrapped dtoa.cc in a namespace to avoid conflict with libmysqlclient
and added it to libutil

Change-Id: I26563038d46ed913348e90072881caffef2ddd19