BUG#31238233: ADOPTING A CLUSTER (ADOPTFROMGR) VIA A SECONDARY MEMBER DISABLES SUPER_READ_ONLY
Adopting a cluster can be performed using any member of the GR group
regardless if it's a primary or a secondary. However, when a secondary
member was used, super_read_only was disabled from that instance when it
should not.

This patch fixes that by ensuring all operations performed during a
cluster adoption are done using the primary member of the group. This
ensures that no GTID inconsistencies occur and that super_read_only is
not incorrectly disabled on secondary members.

Also added regression tests to verify the scenario.

Change-Id: I2d79481b5d66bddb67160da20b195c9b30ba5d30