cmake: add testing component for making testing tarballs
This patch creates a testing component that encloses all the testing binaries
and supporting files to enable running the tests out of the development environment.

In addition, it removes the old dev component generated using the WITH_DEV flag on
the cmake call.

The testing conponent will be generated if the WITH_TESTS flag is set.

Change-Id: I7188a1d552c3bf0b2097930f3eed8055c0fe3e3d