BUG#29255212: CROSS PLATFORM (WIN-LINUX) CLUSTER NOT POSSIBLE DUE LOWER_CASE_TABLE_NAMES
There are some variables that GR requires to be the same on all the members of
the cluster :
https://dev.mysql.com/doc/refman/en/group-replication-options.html and we
while we already validate most of these on the checkInstanceConfiguration
command, there were some that we were missing: lower_case_table_names,
group_replication_gtid_assignment and default_table_encryption.

Unlike the variables that are validated on checkInstanceConfiguration, these
variables have more than 1 value that is compatible with GR and
as such we can only test that they will be the same on all the cluster when we
are adding/rejoining an instance to a cluster.

This patch adds new validations to the cluster.addInstance and
cluster.rejoinInstance commands. These validation prevents adding/rejoining an
instance to the cluster if the value of the lower_case_table_names,
group_replication_gtid_assignment_block_size or default_table_encryption of the
instance are different from the ones on the cluster.

New tests were added for regression and to validate the patch.

Change-Id: I113586c80e068706924eb2b23b5231e68e90116e