In MySQL 8.0.22, WL#9384 introduced changes in Semantics in the sense that a system variable is no longer considered as a "constant" so the validation for the collations is more conservative than in the non-const case.
The Metadata schema 2.0.0 sets several attributes to use the ascii_general_ci
collation which is now detected as incompatible when comparisons with
system variables are performed. Specifically, comparisons of
instances.mysql_server_uuid with @@server_uuid resulted in an error.

This patch fixes that by ensuring the right casts are done in the
AdminAPI queries in use with Metadata schema 2.0.0 or higher.

Change-Id: I6b05137a0ea79d4c3874b8044ef92086b60ff026