BUG#31381697 SHELL DUMP/LOAD: SEGFAULT WHEN RUNNING OUT OF FDS
When dumping data to Object Storage, each file dumped created a
connection to the REST service which was kept alive until the dumper
completed the work which caused to run out of handles.

This patch fixes the issue by ensuring the handles are released as soon
as each file is uploaded.

Change-Id: Ic07dc6d8c404897ac8e03a2fc59b5a5f96d3262e