BUG#31403104 UTIL.DUMPSCHEMAS ALWAYS TRIES TO DUMP USERS (EVEN WHEN SPECIFIED TO NOT DO SO)
When util.dumpSchemas() was executed with 'ocimds' option set to true,
it was checking the compatibility with MDS in SQL which was not
included in the dump.

The compatibility check now takes into account values of other options,
and i.e. if 'dataOnly' is set to true it will not perform the check at
all, if 'users' is set to false, it will not check user accounts, etc.

Change-Id: I295dccb14059408df959b2eac447fd9a509b0560