Fixed defaultCharacterSet handling in loadDump, renamed to characterSet
defaultCharacterSet was not being overriden by the value given as an
option to loadDump(), always using the one specified in the dump
instead. The option was also renamed to characterSet to reflect better
what it does (it overrides the defaults, instead of being the default)

The character set used for loading dumps is the following, in order of
preference:
- characterSet specified as an option to loadDump
- defaultCharacterSet from table.json
- defaultCharacterSet from @.json

Test cases were added to ensure that double-encoding is preserved by
default but fixable by overriding character sets during dump and loading.

Footnote: double-encoding is when utf8mb4 byte data is inserted as latin1
into a utf8mb4 table because of user error. The result is that the data
can only be correctly queried with "set names latin1", even if that's
technically wrong.

Change-Id: I35aba686b771cbaad3b7483d0d96cf154aa155b4