BUG#31420485: MYSQLSH REPLICASET.REMOVEINSTANCE FORCE CHANGED FROM NOTE TO ERROR
The patch for bug 29908657, introduced a regression on the behavior of
<ReplicaSet>.removeInstance():

When removing an instance that is unavailable (offline) from the
Metadata using the 'force' option, the .removeInstance() command prints
an error indicating a connection cannot be established to the instance
when it should not.

This patch fixes that regression by ensuring the command's behavior is
restablished to the right one.

Improved tests cases to cover this and other possible regressions.

Change-Id: I257ac4579d6ec2f7239080ce5b38055983e2e397