Fixed loadDump chunk scheduler when there's just 1 table
When dump contains a single table, only a single
thread was being used even if more were available.

Problem was the proportional chunk scheduling was
not dealing with the case where only 1 table was
available to be loaded. Fix was to change the
default used when no best candidate can be found,
instead of returning nothing, it now returns the
first table in the available list.

Change-Id: If80ef4d96e1ffa6f1b71083eb0b80606f9761167