BUG#31418783: MYSQL SHELL SEGFAULTS WHEN CONNECTED TO PROXYSQL
MySQL Shell, when switching to SQL mode, verifies if ANSI quotes are enabled
by executing the query "select @@sql_mode" and it assumes a result set is
always returned. However, ProxySQL does not return a result set for that query,
when the session is established to its Admin interface.

Even though ProxySQL should forge and return a result set for that
particular query, Shell should not assume the query always returns a result
set.

This patch fixes that by ensuring that if no result set is returned from that
query Shell won't crash and will behave accordingly.

Change-Id: I6e887184b63ef392b18ad863b9557977201df8ed