Fixing REST Service Tests
The REST Service tests assume the host where they are executed contains
a python executable, however in newer hosts the python executable is no
longer available, instead python3 shhould be used.

In addition, test retry_strategy_generic_errors was creating a local
REST server using the same port as the server used on all the tests
when this server was stopped caused the generic server to stop, causing
the rest of the tests to fail.

This patch adds a retry logic to attempt using python3 if using python
failed, and also removes the usage of any local server.

Change-Id: I9b29f637001ffeb06f2ac6ec5737e2a7de9ff0a4