Disabling Reproducible Builds
Some newer Linux distributions have introduced something called
"reproducible builds". In effect the build tools are changed to parse
the changeling entries in the RPM spec file or the Deb changelog file
and use the date string from the most recent entry as "current time".

Even files built are sometimes modified to look like they were built
back in time. Also newer CMake versions honor this and happily returns
a date back in time from STRING(TIMESTAMP somevar). In effect
the build tools acting like a time machine.

This causes trouble as we don't update the changelog entries in
our sources at all or less frequent.

This patch rewrites the Deb changelog with a generic entry with updated
and disable this feature in the RPM spec.

Change-Id: Ifa052fd2805551a2a8742977841dc1dd8dffaf29