Bug#31442908 Bug#31442627 Bug#31442658
This patch introduces new option 'loadIndexes' which can be used
to disable recreation of the indexes after load, which can be useful
in the following scenario:
- Load DDL first with 'loadIndexes' disabled,
- Make some changes to DDL (eg: introduce partitioning),
- Load Data,
- Add secondary indexes.

Indexes recreation will now only start after all the data has been
already loaded for the table.

Loader progress bar changed to display number of threads that are being
indexed.

Change-Id: Ie8aedfa0a31114cc2cd66ab2c25bfe4c5c41fda5