Fixed crash in cluster.status() when called right after a clone process has started in the background:
When the check for clone progress is executed right after clone has started
there's a tiny gap of time on which P_S.clone_progress hasn't been populated
yet. In that scenario, Shell cannot assume there will be data when
quering that table.

This patch fixes a crash due to uninitialized values and in that
scenario simply sets the following attributes of cluster.status() to
"undefined":

- currentStage
- currentStageState
- currentStageProgress

Change-Id: Ib63fac934681b770cc055f7eab9e1b096099b50c