WL#14154 MySQL Shell - Parallel Dump And Loader PAR Flood Extension
This patch introduces an extension into the shell dump and load
functions to enable dumping and loading information into an OCI Bucket
using Preauthenticated Requests (PAR).

Dumper Changes
--------------
A new option 'ociParManifest' has been introduced when dumping to an
OS Bucket that will cause the dumper to generate an Object Read PAR
for each object created. This PAR will be stored into a manifest file
which will contain the PAR information (id, link) and the size for
each object generated on the dump.

A new option 'ociParExpireTime' has been introduced to allow the user
defining the expiration time for the generated PARs. If not specified
the default expiration time will be of 1 week after the dump is created.

Loader Changes
--------------
If the loader receives a PAR to the dump manifest (@.manifest.json) on
the URL parameter, it will cause the load to be executed using the PARs
contained on the manifest.

Additional Changes
------------------
The code that ensures conpatibility when uploading data to MDS has been
updated to:
- Use a whitelist of grants to determine if a grant is valid or not for MDS.
  This whitelist is the grants that are given to the admin user when a
  DB System is deployed.
- Removed the 'strip_admin_role' compatibility option as it is no longer
  needed with the new logic described above.
- Removed the FLUSH GRANTS statement at the end of the users file in a dump
  as such statement is not required and can't be executed when importing
  data to MDS.

Change-Id: Ibd451a9b966020e6aa19fc821365ebfe56191d19