Fixed table scheduling when loading dumps
Loader was scheduling chunks from the same table even when there were
different tables waiting to be loaded in some cases.

There was an int overflow when handling large tables that was breaking
proportional scheduling. Additionally, changed the scheduling algorithm
to explicitly and separately handle the case where there are more
threads than tables left to be loaded.

- Added a mock test to simulate loading a dump from a list of files and
metadata files.

- Improved query handling in mock session/results

Change-Id: I0248194af39a4a438b00315f88d44f455cabc0fe