BUG#31606223 BUG#31605287 PROGRESS FILE LINKED TO A PAR OBJECT WITH READ/WRITE ACCESS THROWS AN EXCEPTION
When the util.loadDump() method was used to load a dump created with PAR
metadata, and the 'progressFile' option was set to a PAR file which
existed, but contained no data (zero-length file), an exception was
thrown.

This issue is now fixed, previously the existence of a PAR file was
checked using a GET request with a Range header set to 0-0, which for
an empty file results in 416 Range Not Satisfiable error. This has been
replaced with a HEAD request.

Additionally, if 'progressFile' was set to an empty, existing file, the
loader would still assume that the previous attempt to load the dump was
interrupted and disable some of its checks i.e. the one for existing
schemas/tables, effectively overwriting the whole instance. This is
also fixed, if an empty progress file is given, the loader will
correctly identify the dump as still pending (and not interrupted).

Change-Id: I13ad8fca9827587a5c11e41c485e922903715708