BUG#31605353 RESETPROGRESS SUGGESTED BUT NOT WORKING - ERROR: UTIL.LOADDUMP: CANNOT SETUP SIG
When loading a dump using a PAR to a progress file combined with the
'resetProgress' option an exception is thrown.

If the 'resetProgress' option is used loader will delete the progress
file, in case of PARs DELETE method is not supported, so the issue was
fixed by truncating the progress file.

Change-Id: I4d6450e7a6a90ff7d6a6555c59519b5a275aff42